precommit fix on new submodules
