+++ docs/api/selfeeg.ssl.base.evaluate_loss.rst
@@ -1,6 +1,6 @@
 evaluate_loss
 =============
 
-.. currentmodule:: selfeeg.ssl.ssl
+.. currentmodule:: selfeeg.ssl.base
 
 .. autofunction:: evaluate_loss
