+++ docs/api/selfeeg.ssl.base.fine_tune.rst
@@ -1,6 +1,6 @@
 fine_tune
 =========
 
-.. currentmodule:: selfeeg.ssl.ssl
+.. currentmodule:: selfeeg.ssl.base
 
 .. autofunction:: fine_tune
